+++ src/lib/deleteHandler.js
@@ -26,7 +26,7 @@ const AUTH = require('./sharedConstants').AUTH;
 const logger = new Logger(module);
 
 // This is an ordered list - objects will be deleted in this order
-const DELETABLE_CLASSES = ['DeviceGroup', 'DNS_Resolver', 'Route', 'SelfIp', 'VLAN', 'Trunk', 'RouteDomain', 'RemoteAuthRole', 'ManagementRoute', 'Tunnel', 'RoutingAsPath', 'RoutingPrefixList', 'GSLBServer', 'GSLBDataCenter'];
+const DELETABLE_CLASSES = ['DeviceGroup', 'DNS_Resolver', 'Route', 'SelfIp', 'VLAN', 'Trunk', 'RouteDomain', 'RemoteAuthRole', 'ManagementRoute', 'Tunnel', 'RoutingAsPath', 'RoutingPrefixList', 'GSLBServer', 'GSLBDataCenter', 'GSLBMonitor'];
 
 const READ_ONLY_DEVICE_GROUPS = ['device_trust_group', 'gtm', 'datasync-global-dg', 'dos-global-dg'];
 
@@ -81,6 +81,10 @@ class DeleteHandler {
                 },
                 DNS_Resolver: {
                     'f5-aws-dns': {} // referenced by Security Offbox Service f5-tap-ingress-aws-global
+                },
+                GSLBMonitor: { // BIG-IP has a number of default monitors, that cannot be removed
+                    http: {},
+                    http_head_f5: {}
                 }
             };
 
@@ -105,6 +109,12 @@ class DeleteHandler {
                         const partition = this.state.currentConfig.Common.Route[itemToDelete].localOnly
                             ? 'LOCAL_ONLY' : 'Common';
                         const path = `${PATHS.Route}/~${partition}~${itemToDelete}`;
+                        classPromises.push(this.bigIp.delete(path, null, null, cloudUtil.NO_RETRY));
+                    } else if (deleteableClass === 'GSLBMonitor'
+                        && !isRetainedItem(deleteableClass, itemToDelete)) {
+                        // GSLB Monitors have their monitor type as part of the path instead of a property
+                        const currMon = this.state.currentConfig.Common.GSLBMonitor[itemToDelete];
+                        const path = `${PATHS.GSLBMonitor}/${currMon.monitorType}/~Common~${itemToDelete}`;
                         classPromises.push(this.bigIp.delete(path, null, null, cloudUtil.NO_RETRY));
                     } else if (!isRetainedItem(deleteableClass, itemToDelete)) {
                         const commonPrefix = deleteableClass === 'Trunk' ? '' : '~Common~';